INGK-1054 Move XMLBase to dictionary.py and use it to dictionary and configuration file classes
